Разбивает файл утилит на два, добавляет абстрактный класс и его использование компонентами
