+++ src/components/info.js
@@ -1,5 +1,5 @@
 import {MONTH_NAMES} from "../const.js";
-import {createElement} from "../utils.js";
+import AbstractComponent from "./abstract-component.js";
 
 const createTripInfoTemplate = (events) => {
   if (!events.length) {
@@ -24,25 +24,14 @@ const createTripInfoTemplate = (events) => {
   );
 };
 
-export default class Info {
+export default class Info extends AbstractComponent {
   constructor(events) {
-    this._element = null;
+    super();
+
     this._events = events;
   }
 
   getTemplate() {
     return createTripInfoTemplate(this._events);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
