+++ src/components/no-events.js
@@ -1,27 +1,11 @@
-import {createElement} from "../utils.js";
+import AbstractComponent from "./abstract-component.js";
 
 const createNoEventsTemplate = () => {
   return (`<p class="trip-events__msg">Click New Event to create your first point</p>`);
 };
 
-export default class NoEvents {
-  constructor() {
-    this._element = null;
-  }
-
+export default class NoEvents extends AbstractComponent {
   getTemplate() {
     return createNoEventsTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
