+++ src/utils/common.js
@@ -1,10 +1,3 @@
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`,
-  BEFOREBEGIN: `beforebegin`,
-  AFTEREND: `afterend`
-};
-
 export const getRandomArrayElement = (array) =>
   array[Math.floor(Math.random() * array.length)];
 
@@ -59,26 +52,3 @@ export const formatDuration = (date) => {
 
 export const capitalizeFirstLetter = (string) =>
   string.charAt(0).toUpperCase() + string.slice(1);
-
-export const createElement = (elementTemplate) => {
-  const tempElement = document.createElement(`div`);
-  tempElement.innerHTML = elementTemplate;
-  return tempElement.firstChild;
-};
-
-export const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-    case RenderPosition.BEFOREBEGIN:
-      container.before(element);
-      break;
-    case RenderPosition.AFTEREND:
-      container.after(element);
-      break;
-  }
-};
